T#767: Extract prowl routes to src/prowl/routes.ts
Phase 1.2 of Library #102 modularization. Moves 12 route handlers (tasks CRUD,
checklists, toggle, notify-test) + ALLOWED_PROWL constants from server.ts to
src/prowl/routes.ts. Table creation stays in server.ts. No logic changes.
